requirements.txt
# Jumpy Bird - Production Dependencies
# Generated for Render deployment

Flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
